requirements.txt
coverage==5.0.2
Django==2.2.7
django-admin==2.0.0
django-cors-headers==3.2.0
django-excel-response2==2.0.8
django-rest-auth==0.9.5
django-six==1.0.4
djangorestframework==3.10.3
Faker==3.0.0
mock==3.0.5
mysqlclient==1.4.5
numpy==1.18.0
opencv-python==4.1.2.30
Pillow==6.2.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.10.3
pytz==2019.3
screen==1.0.1
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
xlwt==1.3.0
